Move main UI into common component used by both normal mode and Deforum mode.
