+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sd-parseq",
-  "version": "0.0.1",
+  "version": "0.0.3",
   "private": true,
   "dependencies": {
     "@ahooksjs/use-url-state": "^3.5.0",
